feat: Enhance VaM Resource Browser with new features and improvements - Updated plugin build script error messages for clarity - Added new script for handling hair reference meshes - Implemented graft boundary transfer functionality - Improved geometry preview handling in UE - Added support for PythonScriptPlugin in the project configuration - Updated documentation for Stage 04 with detailed usage instructions - Enhanced material handling and error checking in source materials - Added new preview request handling script

+++ Scripts/vam_materials.py
@@ -242,7 +242,7 @@ def build_material_ir(plan,ir,preview,out):
                 for m in affected:m['parameters'][PARAMS[key]]=color(value) if isinstance(value,dict) else float(value)
             elif key in ('hideMaterial','renderQueue'):
                 used.add(key)
-                for m in affected:m['render_state']['hidden' if key=='hideMaterial' else 'render_queue']=str(value).lower()=='true' if key=='hideMaterial' else int(value)
+                for m in affected:m['render_state']['hidden' if key=='hideMaterial' else 'render_queue']=str(value).lower()=='true' if key=='hideMaterial' else float(value)
             elif key.startswith('customTexture_'):
                 used.add(key);prop=key[len('customTexture'):]
                 if not value:
